requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.1
python-multipart==0.0.9
pdfplumber==0.11.4
PyPDF2==3.0.1
python-docx==1.1.2
pytesseract==0.3.13
Pillow==10.4.0
openai==1.51.2
httpx==0.27.2
tenacity==9.0.0
rapidfuzz==3.9.7
aiofiles==23.2.1
deep-translator==1.11.4
huggingface_hub==0.24.1

